Cleanup write4

+++ encode_map.go
@@ -127,7 +127,7 @@ func (e *Encoder) EncodeMapLen(l int) error {
 	if l < 65536 {
 		return e.write2(codes.Map16, uint64(l))
 	}
-	return e.write4(codes.Map32, uint64(l))
+	return e.write4(codes.Map32, uint32(l))
 }
 
 func encodeStructValue(e *Encoder, strct reflect.Value) error {
